# By Michael Groeber (9) and Michael Jackson (2)
# Via Michael Groeber (4) and Michael Jackson (2)
* 'develop' of ssh://github.com/dream3d/DREAM3D:
  adding a Modified Lambert Projection Test case using images supplied by Marc DeGraef
  fixing neighborlist.hpp implementation
  trying to update the neighborlist.hpp file
  more work in orientationOps classes
  working on TSL IPF coloring
  Fixing compile error when compiling non parallel codes that have a parallel option,
  adding a constant
  doing some work on the mic file origin stuff
  changing the burn algorithm in segment grains to reduce the memory needed to hold the running list of voxels in the current grain
  working on some 64bit compatibility issues
  adding ability to read mic file in read orientation data, which also allows for the viewing of the mic in the import filter when setting the reference frame(s)

+++ Source/DREAM3DLib/ReconstructionFilters/ScalarSegmentGrains.cpp
@@ -394,7 +394,7 @@ void ScalarSegmentGrains::execute()
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-int ScalarSegmentGrains::getSeed(size_t gnum)
+int64_t ScalarSegmentGrains::getSeed(size_t gnum)
 {
   setErrorCondition(0);
   VoxelDataContainer* m = getVoxelDataContainer();
@@ -410,17 +410,17 @@ int ScalarSegmentGrains::getSeed(size_t gnum)
   int64_t totalPoints = m->getTotalPoints();
 
   DREAM3D_RANDOMNG_NEW()
-  int seed = -1;
-  int randpoint = 0;
+  int64_t seed = -1;
+  int64_t randpoint = 0;
 
   // Pre-calculate some constants
   int64_t totalPMinus1 = totalPoints - 1;
 
-  int counter = 0;
-  randpoint = int(float(rg.genrand_res53()) * float(totalPMinus1));
+  int64_t counter = 0;
+  randpoint = int64_t(float(rg.genrand_res53()) * float(totalPMinus1));
   while (seed == -1 && counter < totalPoints)
   {
-      if (randpoint > totalPMinus1) randpoint = static_cast<int>( randpoint - totalPoints );
+      if (randpoint > totalPMinus1) randpoint = static_cast<int64_t>( randpoint - totalPoints );
       if (m_GrainIds[randpoint] == 0) seed = randpoint;
       randpoint++;
       counter++;
@@ -437,7 +437,7 @@ int ScalarSegmentGrains::getSeed(size_t gnum)
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-bool ScalarSegmentGrains::determineGrouping(int referencepoint, int neighborpoint, size_t gnum)
+bool ScalarSegmentGrains::determineGrouping(int64_t referencepoint, int64_t neighborpoint, size_t gnum)
 {
 
   if(m_GrainIds[neighborpoint] > 0) { return false; }
